[S07P31D207-270] feat-be : 한번에  컨테이너 삭제 test

See merge request s07-final/S07P31D207!516

+++ goldenCrow/src/main/java/com/example/goldencrow/compile/SchedulerService.java
@@ -12,7 +12,7 @@ public class SchedulerService {
     @Autowired
     private CompileService compileService;
 
-    @Scheduled(cron = "0 10 * * * *")
+    @Scheduled(cron = "0 20 * * * *")
     public void run() {
         System.out.println("hi i'm working");
         // 모든 컨테이너 닫기
@@ -26,9 +26,13 @@ public class SchedulerService {
         * docker rmi $(docker images -f "dangling=true" -q)
         * */
         String filteringName = "crowstudio_";
-        String[] stopCmd = {"/bin/sh", "-c", "containerlist=", "$(docker", "container", "ls", "--filter='name="+ filteringName +"'", "-q);", "docker", "stop", "$containerlist"};
+        String[] containerCmd = {"docker", "container", "ls", "--filter='name="+ filteringName +"'", "-q"};
+        String containerList = compileService.resultString(containerCmd);
+        if (containerList == null) { return; }
+        String[] stopCmd = {"docker", "stop", containerList};
         System.out.println("docker stop 시작 !");
         compileService.resultString(stopCmd);
+
         System.out.println("docker stop 됐다 !");
         String[] rmImgCmd = {"/bin/sh", "-c", "docker", "rmi", "$(docker", "images", filteringName +"*", "-q)"};
         System.out.println("docker images 삭제 시작 !");
